+++ src/calendar.js
@@ -2478,6 +2478,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 buildToolbarButton( titleBar, "ib-arrow-right-full-line", _options.jumpToDateTitle, function() {
                     showJumpToDateDialog();
                 } );
+
+                titleBar.appendChild( createElement( "div", "left-divider-line" ) );
     
                 buildToolbarButton( titleBar, "ib-refresh", _options.refreshTooltipText, function() {
                     refreshViews( true, true );
@@ -2963,6 +2965,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 buildToolbarButton( titleBar, "ib-arrow-right-full-line", _options.jumpToDateTitle, function() {
                     showJumpToDateDialog();
                 } );
+
+                titleBar.appendChild( createElement( "div", "left-divider-line" ) );
 
                 buildToolbarButton( titleBar, "ib-refresh", _options.refreshTooltipText, function() {
                     refreshViews( true, true );
@@ -3587,6 +3591,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
             buildToolbarButton( _element_View_FullMonth_TitleBar, "ib-arrow-right-full-line", _options.jumpToDateTitle, function() {
                 showJumpToDateDialog();
             } );
+
+            _element_View_FullMonth_TitleBar.appendChild( createElement( "div", "left-divider-line" ) );
 
             buildToolbarButton( _element_View_FullMonth_TitleBar, "ib-refresh", _options.refreshTooltipText, function() {
                 refreshViews( true, true );
@@ -4577,6 +4583,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     showJumpToDateDialog();
                 } );
 
+                titleBar.appendChild( createElement( "div", "left-divider-line" ) );
+
                 buildToolbarButton( titleBar, "ib-refresh", _options.refreshTooltipText, function() {
                     refreshViews( true, true );
                 } );
@@ -5262,6 +5270,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 buildToolbarButton( titleBar, "ib-arrow-right-full-line", _options.jumpToDateTitle, function() {
                     showJumpToDateDialog();
                 } );
+
+                titleBar.appendChild( createElement( "div", "left-divider-line" ) );
 
                 buildToolbarButton( titleBar, "ib-refresh", _options.refreshTooltipText, function() {
                     refreshViews( true, true );